+++ app/app.py
@@ -1,9 +1,9 @@
 from flask import Flask
 from flask_cors import CORS
-from views.contrato import app as contrato
+from views.contract import app as contract
 
 app = Flask(__name__)
-app.register_blueprint(contrato)
+app.register_blueprint(contract)
 CORS(app, automatic_options=True)
