+++ app/utils/constants.py
@@ -1,9 +1,9 @@
 DATABASE_CONFIG = {
-    "DB_HOST": "serviceName",
+    "DB_HOST": "contract-mongodb",
     "DB_PORT": 27017,
     "DB_USERNAME": "admin",
     "DB_PASSWORD": "password",
     "authSource": "",
     "authMechanism": "SCRAM-SHA-256",
-    "collection": "nameCollection"
+    "collection": "contract"
 }
