+++ app/views/contract.py
@@ -1,0 +1,11 @@
+from flask import Blueprint, request, jsonify
+from flask_cors import CORS
+import controllers.contract as contract
+
+
+app = Blueprint('contract', __name__)
+CORS(app)
+
+@app.route("/contract", methods=["POST"])
+def contrato():
+    return contract.register_contract_request(request.json)
